backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0
